Merge changes uint16_t to uint64_t where it was used in machine's
code.
This change is essential as it does away with the 65KB limit the
uint16_t imposed on addresses in and sizes of bytecode.

Fixes #94.

+++ include/viua/cg/assembler/assembler.h
@@ -37,17 +37,17 @@ namespace assembler {
     }
 
     namespace verify {
-        std::string functionCallsAreDefined(const std::vector<std::string>& lines, const std::map<unsigned, unsigned>&, const std::vector<std::string>& function_names, const std::vector<std::string>& function_signatures);
-        std::string frameBalance(const std::vector<std::string>& lines, const std::map<unsigned, unsigned>&);
-        std::string callableCreations(const std::vector<std::string>& lines, const std::map<unsigned, unsigned>&, const std::vector<std::string>& function_names, const std::vector<std::string>& function_signatures);
-        std::string ressInstructions(const std::vector<std::string>& lines, const std::map<unsigned, unsigned>&, bool as_lib);
+        std::string functionCallsAreDefined(const std::vector<std::string>& lines, const std::map<long unsigned, long unsigned>&, const std::vector<std::string>& function_names, const std::vector<std::string>& function_signatures);
+        std::string frameBalance(const std::vector<std::string>& lines, const std::map<long unsigned, long unsigned>&);
+        std::string callableCreations(const std::vector<std::string>& lines, const std::map<long unsigned, long unsigned>&, const std::vector<std::string>& function_names, const std::vector<std::string>& function_signatures);
+        std::string ressInstructions(const std::vector<std::string>& lines, const std::map<long unsigned, long unsigned>&, bool as_lib);
         std::string functionBodiesAreNonempty(const std::vector<std::string>& lines, std::map<std::string, std::vector<std::string> >& functions);
-        std::string blockTries(const std::vector<std::string>& lines, const std::map<unsigned, unsigned>&, const std::vector<std::string>& block_names, const std::vector<std::string>& block_signatures);
+        std::string blockTries(const std::vector<std::string>& lines, const std::map<long unsigned, long unsigned>&, const std::vector<std::string>& block_names, const std::vector<std::string>& block_signatures);
         std::string blockBodiesEndWithLeave(const std::vector<std::string>& lines, std::map<std::string, std::pair<bool, std::vector<std::string> > >& blocks);
         std::string mainFunctionDoesNotEndWithHalt(std::map<std::string, std::vector<std::string> >& functions);
 
-        std::string directives(const std::vector<std::string>& lines, const std::map<unsigned, unsigned>&);
-        std::string instructions(const std::vector<std::string>& lines, const std::map<unsigned, unsigned>&);
+        std::string directives(const std::vector<std::string>& lines, const std::map<long unsigned, long unsigned>&);
+        std::string instructions(const std::vector<std::string>& lines, const std::map<long unsigned, long unsigned>&);
     }
 }
